# lab_wizard/wizard/frontend/src/lib/api/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "LaunchStatus": {
    "description": "Whether the project is running from here: ``idle``, ``starting``, ``running`` or ``ended``.",
    "properties": {
     "exit_code": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Exit Code"
     },
     "launch_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Launch Id"
     },
     "log": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Log"
     },
     "log_file": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Log File"
     },
     "pid": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pid"
     },
     "run_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run Id"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "state": {
      "enum": [
       "idle",
       "starting",
       "running",
       "ended"
      ],
      "title": "State",
      "type": "string"
     }
    },
    "required": [
     "state",
     "launch_id",
     "pid",
     "started_at",
     "run_id",
     "exit_code",
     "log",
     "log_file"
    ],
    "title": "LaunchStatus",
    "type": "object"
   },
...
   "OutputsConfig-Output": {
    "description": "What a run produces besides its record in the lab database, which is always written.",
    "properties": {
     "files": {
      "default": true,
      "title": "Files",
      "type": "boolean"
     },
     "live_plot": {
      "default": "none",
      "enum": [
       "none",
       "window",
       "web"
      ],
      "title": "Live Plot",
      "type": "string"
     },
     "plot": {
      "default": "",
      "title": "Plot",
      "type": "string"
     }
    },
    "required": [
     "files",
     "live_plot",
     "plot"
    ],
    "title": "OutputsConfig",
    "type": "object"
   },
   "ProjectSettings": {
    "description": "A project's settings, as the Run page edits them: fields, and the YAML they come from.",
    "properties": {
     "kind": {
      "enum": [
       "procedure",
       "custom"
      ],
      "title": "Kind",
      "type": "string"
     },
     "measurement": {
      "title": "Measurement",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "outputs": {
      "$ref": "#/components/schemas/OutputsConfig-Output"
     },
     "params": {
      "additionalProperties": true,
      "title": "Params",
      "type": "object"
     },
     "params_schema": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Params Schema"
     },
     "path": {
      "title": "Path",
      "type": "string"
     },
     "run": {
      "$ref": "#/components/schemas/RunConfig"
     },
     "setup_file": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Setup File"
     },
     "yaml": {
      "title": "Yaml",
      "type": "string"
     }
    },
    "required": [
     "name",
     "path",
     "measurement",
     "kind",
     "setup_file",
     "yaml",
     "run",
     "params",
     "outputs",
     "params_schema"
    ],
    "title": "ProjectSettings",
    "type": "object"
   },
   "RunConfig": {
    "description": "Who and what a run is about. Read at the start of every run and recorded with it.",
    "properties": {
     "device": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "The device under test, by its name in the lab database",
      "title": "Device"
     },
     "metadata": {
      "additionalProperties": true,
      "title": "Metadata",
      "type": "object"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "operator": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Operator"
     }
    },
    "required": [
     "device",
     "operator",
     "notes",
     "metadata"
    ],
    "title": "RunConfig",
    "type": "object"
   },
...
   "SettingsBody": {
    "description": "The whole YAML, or any of the sections.",
    "properties": {
     "outputs": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Outputs"
     },
     "params": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Params"
     },
     "run": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run"
     },
     "yaml": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Yaml"
     }
    },
    "title": "SettingsBody",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
   "_DefinitionBody": {
    "properties": {
     "definition": {
      "additionalProperties": true,
      "title": "Definition",
      "type": "object"
     }
    },
    "required": [
     "definition"
    ],
    "title": "_DefinitionBody",
    "type": "object"
   },
...
   "_PresetBody": {
    "properties": {
     "values": {
      "additionalProperties": true,
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "values"
    ],
    "title": "_PresetBody",
    "type": "object"
   },
...
   "_SavePlotBody": {
    "properties": {
     "plot": {
      "additionalProperties": true,
      "title": "Plot",
      "type": "object"
     },
     "replace": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Replace"
     }
    },
    "required": [
     "plot"
    ],
    "title": "_SavePlotBody",
    "type": "object"
   },
...
   "_YamlBody": {
    "properties": {
     "yaml": {
      "title": "Yaml",
      "type": "string"
     }
    },
    "required": [
     "yaml"
    ],
    "title": "_YamlBody",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/procedures/check": {
   "post": {
    "description": "Problems with a definition being edited, each with its path, and its Python.",
    "operationId": "api_procedure_check_api_procedures_check_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_DefinitionBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Check"
   }
  },
  "/api/procedures/from-yaml": {
   "post": {
    "operationId": "api_procedure_from_yaml_api_procedures_from_yaml_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_YamlBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure From Yaml"
   }
  },
  "/api/procedures/to-yaml": {
   "post": {
    "operationId": "api_procedure_to_yaml_api_procedures_to_yaml_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_DefinitionBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure To Yaml"
   }
  },
  "/api/procedures/{name}": {
   "delete": {
    "description": "Delete this workspace's copy. Deleting an override brings the built-in back.",
    "operationId": "api_procedure_delete_api_procedures__name__delete",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Delete"
   },
   "get": {
    "operationId": "api_procedure_detail_api_procedures__name__get",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Detail"
   },
   "put": {
    "description": "Save to this workspace. A built-in's name saves a workspace override of it.",
    "operationId": "api_procedure_save_api_procedures__name__put",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_DefinitionBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Save"
   }
  },
  "/api/procedures/{name}/plots": {
   "post": {
    "description": "Save a plot built on the Data page into its procedure.",
    "operationId": "api_procedure_save_plot_api_procedures__name__plots_post",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_SavePlotBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Save Plot"
   }
  },
  "/api/procedures/{name}/presets": {
   "get": {
    "operationId": "api_procedure_presets_api_procedures__name__presets_get",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Presets"
   }
  },
  "/api/procedures/{name}/presets/{preset}": {
   "delete": {
    "operationId": "api_procedure_preset_delete_api_procedures__name__presets__preset__delete",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "preset",
      "required": true,
      "schema": {
       "title": "Preset",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Preset Delete"
   },
   "put": {
    "operationId": "api_procedure_preset_save_api_procedures__name__presets__preset__put",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "preset",
      "required": true,
      "schema": {
       "title": "Preset",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_PresetBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Procedure Preset Save"
   }
  },
...
  "/api/projects/{name}/launch": {
   "get": {
    "operationId": "api_project_launch_status_api_projects__name__launch_get",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LaunchStatus"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Project Launch Status"
   },
   "post": {
    "description": "Run the project's setup file, as its own process.",
    "operationId": "api_project_launch_api_projects__name__launch_post",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LaunchStatus"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Project Launch"
   }
  },
  "/api/projects/{name}/settings": {
   "get": {
    "description": "The project's run details, params and outputs, as fields and as its YAML.",
    "operationId": "api_project_settings_api_projects__name__settings_get",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectSettings"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Project Settings"
   },
   "put": {
    "description": "Save ``{\"yaml\"}``, or any of ``{\"run\", \"params\", \"outputs\"}``; 422 lists every problem.",
    "operationId": "api_project_settings_save_api_projects__name__settings_put",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SettingsBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectSettings"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Project Settings Save"
   }
  },
  "/api/projects/{name}/stop": {
   "post": {
    "description": "Ctrl-C the run: it aborts and puts its instruments in their safe state.",
    "operationId": "api_project_stop_api_projects__name__stop_post",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LaunchStatus"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Project Stop"
   }
  },
...
 }
}